+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 2.0.1 Release
 
+### New features
+
+- Migration script to convert backup files taken by OS cronjob into k8s compatible format [#13099](https://github.com/opencrvs/opencrvs-core/issues/13099)
+
 ### Bug fixes
 
 - Always restart the Kubernetes self-hosted runner during deployment to ensure the latest runner image and configuration changes are applied. [#332](https://github.com/opencrvs/infrastructure/pull/332)